+++ BxInputController/Sources/Protocols/BxInputStaticHeight.swift
@@ -1,0 +1,15 @@
+//
+//  BxInputStaticHeight.swift
+//  BxInputController
+//
+//  Created by Sergey Balalaev on 23/01/17.
+//  Copyright © 2017 Byterix. All rights reserved.
+//
+
+import UIKit
+
+protocol BxInputStaticHeight {
+    
+    var height : CGFloat {get}
+    
+}