+++ CHANGELOG.md
@@ -9,6 +9,7 @@ ByteriX, 2017. All right reserved.
 ## 0.9.19 (20.01.2017)
 ##### Bug fixing
 * refactoring: extract protocol BxInputString for all
+* abstraction variant type for component
 
 ## 0.9.18 (19.01.2017)
 ##### Bug fixing